fix user-agent and refactor code for interceptor for request header
