+++ net/common/src/main/java/de/danoeh/antennapod/net/common/RequestHeaderIntercepter.java
@@ -1,0 +1,41 @@
+package de.danoeh.antennapod.net.common;
+
+import androidx.annotation.NonNull;
+
+import okhttp3.Interceptor;
+import okhttp3.Response;
+
+import java.io.IOException;
+import java.util.Locale;
+
+public class RequestHeaderIntercepter implements Interceptor {
+    public static String ACCEPT_ENCODING = "gzip, deflate, br";
+   // public static String ACCEPT_ENCODING = "gzip";
+
+
+    @Override
+    public Response intercept(Chain chain) throws IOException {
+        String existingEncoding = chain.request().header("Accept-Encoding");
+        String newEncodingValue = getEncodeString(existingEncoding);
+        return chain.proceed(chain.request().newBuilder()
+                .header("Accept-Encoding", newEncodingValue)
+                .build());
+    }
+    @NonNull
+    private static String getEncodeString(String existingEncoding) {
+        String newEncodingValue;
+        if (existingEncoding == null || existingEncoding.isEmpty()) {
+            newEncodingValue = ACCEPT_ENCODING;
+        } else {
+            StringBuilder encodingBuilder = new StringBuilder(existingEncoding);
+            if (!existingEncoding.toLowerCase(Locale.ROOT).contains("deflate")) {
+              encodingBuilder.append(", deflate");
+            }
+            if (!existingEncoding.toLowerCase(Locale.ROOT).contains("br")) {
+              encodingBuilder.append(", br");
+            }
+            newEncodingValue = encodingBuilder.toString();
+        }
+        return newEncodingValue;
+    }
+}
